Initial commit
The project approximates the value of pi by utilizing monte carlo. This is done by solving for the ratio of N points inside a circle circumscribed by a square of equal radii to the number of points outside the circle.

This project contains two implementations of monte carlo in python. monteCarlo.py is executed in serial with a tkinter window for visualization. monteCarloThreaded.py is executed threaded and displays its output solely in the terminal.